impr: re-organize parser modules, remove peg parser
Add a grammar module as a submodule of the proper-API providing parser module.
Re-organize parser rules into separate submodules of the grammar module.
Remove PEG parser and grammar.
