+++ Cargo.toml
@@ -13,8 +13,6 @@ path = "src/main.rs"
 log = "*"
 nom = "*"
 num = "*"
-peg = "*"
-peg-syntax-ext = "*"
 rand = "*"
 regex = "*"
 rustc-serialize = "*"
