+++ src/command/fetch.rs
@@ -11,15 +11,14 @@ use super::store::StoreName::Add;
 
 /// Take the rest of the arguments provided by the client and parse them into a
 /// Command object with command::fetch.
-pub fn fetch(args: Split<char>) -> Result<Command, parser::ParseError> {
+pub fn fetch(args: Split<char>) -> Result<Command, ()> {
     let mut cmd = "FETCH".to_string();
     for arg in args {
         cmd.push(' ');
         cmd.push_str(arg);
     }
 
-    // Parse the command with the PEG parser.
-   parser::fetch(&cmd[..])
+    parser::fetch(cmd.as_bytes())
 }
 
 /// Perform the fetch operation on each sequence number indicated and return
