--- src/grammar.rustpeg
@@ -1,265 +1,0 @@
-use std::ascii::AsciiExt;
-
-use command::{
-    Attribute,
-    Command
-};
-use command::Attribute::{
-    Body,
-    BodyPeek,
-    BodySection,
-    BodyStructure,
-    Envelope,
-    Flags,
-    InternalDate,
-    RFC822,
-    UID
-};
-use command::CommandType::Fetch;
-use command::RFC822Attribute::{
-    AllRFC822,
-    HeaderRFC822,
-    SizeRFC822,
-    TextRFC822
-};
-use command::sequence_set::SequenceItem;
-use command::sequence_set::SequenceItem::{
-    Number,
-    Range,
-    Wildcard
-};
-
-use mime::{BodySectionType, Msgtext};
-use mime::BodySectionType::{
-    AllSection,
-    MsgtextSection,
-    PartSection
-};
-use mime::Msgtext::{
-    HeaderMsgtext,
-    HeaderFieldsMsgtext,
-    HeaderFieldsNotMsgtext,
-    TextMsgtext,
-    MimeMsgtext
-};
-
-/* FETCH command rules */
-
-pub fetch -> Command
-    = "FETCH"i SP set:sequence_set SP attrs:(
-    "(" a:fetch_att b:(SP f:fetch_att { f })* ")"
-    {
-        let mut attrs = vec![a];
-        for attr in b.into_iter() {
-            attrs.push(attr);
-        }
-        attrs
-    }
-    / attr:fetch_att { vec![attr] }
-    / "ALL"i { vec![Flags, InternalDate, RFC822(SizeRFC822), Envelope] }
-    / "FULL"i { vec![Flags, InternalDate, RFC822(SizeRFC822), Envelope, Body] }
-    / "FAST"i { vec![Flags, InternalDate, RFC822(SizeRFC822)] })
-    { Command::new(Fetch, set, attrs) }
-
-fetch_att -> Attribute
-    = "ENVELOPE"i { Envelope }
-    / "FLAGS"i { Flags }
-    / "INTERNALDATE"i { InternalDate }
-    / "RFC"i "822" sub_attr:(".HEADER" { HeaderRFC822 } / ".SIZE" { SizeRFC822 } / ".TEXT" { TextRFC822 })? { RFC822(sub_attr.unwrap_or(AllRFC822)) }
-    / "UID"i { UID }
-    / "BODY"i "." "PEEK"i section:section octets:("<" first_octet:number "." last_octet:nz_number ">" { (first_octet, last_octet) })? { BodyPeek(section, octets) }
-    / "BODY"i section:section octets:("<" first_octet:number "." last_octet:nz_number ">" { (first_octet, last_octet) })? { BodySection(section, octets) }
-    / "BODY"i sub_attr:("STRUCTURE" { BodyStructure })? { sub_attr.unwrap_or(Body) }
-
-section -> BodySectionType
-    = "[" section:(section_spec)? "]"
-    { section.unwrap_or(AllSection) }
-
-section_spec -> BodySectionType
-    = a:section_msgtext { MsgtextSection(a) }
-    / (a:section_part b:("." s:section_text { s })? { PartSection(a, b) })
-
-section_msgtext -> Msgtext
-    = "HEADER"i "." "FIELDS"i not:("." "NOT"i)? SP headers:header_list {
-            match not {
-                Some(_) => {
-                    HeaderFieldsNotMsgtext(headers)
-                },
-                None => {
-                    HeaderFieldsMsgtext(headers)
-                }
-            }
-        }
-    / "HEADER"i { HeaderMsgtext }
-    / "TEXT"i { TextMsgtext }
-    /* top-level or MESSAGE/RFC822 part */
-
-header_list -> Vec<String>
-    = "(" a:header_fld_name b:(SP h:header_fld_name { h })* ")"
-    {
-        let mut fields = vec![a];
-        for field in b.into_iter() {
-            fields.push(field);
-        }
-        fields
-    }
-
-header_fld_name -> String
-    = a:astring
-    { (&a[..]).to_ascii_uppercase().to_string() }
-
-section_part -> Vec<usize>
-    = a:nz_number b:("." n:nz_number { n })*
-    {
-        let mut parts = vec![a];
-        for part in b.into_iter() {
-            parts.push(part);
-        }
-        parts
-    }
-    /* body part nesting */
-
-section_text -> Msgtext
-    = section_msgtext / "MIME"i { MimeMsgtext }
-    /* text other than actual body part (headers, etc.) */
-
-astring -> String
-    = (s:$(ASTRING_CHAR+) { s.to_string() }) / string
-
-ASTRING_CHAR
-    = ATOM_CHAR / resp_specials
-
-ATOM_CHAR
-    = (!atom_specials CHAR)
-    /* any CHAR except atom_specials */
-
-atom_specials
-    = "(" / ")" / "{" / SP / CTL / list_wildcards
-    / quoted_specials / resp_specials
-
-CTL
-    = [\x00-\x1F] / [\x7F]
-    /* control characters */
-
-list_wildcards
-    = "%" / "*"
-
-quoted_specials
-    = DQUOTE / "\\"
-
-resp_specials
-    = "]"
-
-string -> String
-    = quoted / literal
-
-quoted -> String
-    = DQUOTE quoted_string:(c:$(QUOTED_CHAR*) { c.to_string() }) DQUOTE
-    { quoted_string }
-
-literal -> String
-    = "{" number "}" CRLF (c:$(CHAR8*) { c.to_string() })
-    /*
-     * the "number" is used to indicate the number of octets.
-     */
-
-QUOTED_CHAR
-    = (!quoted_specials TEXT_CHAR)
-    / "\\" quoted_specials
-    /* any TEXT_CHAR except quoted_specials */
-
-TEXT_CHAR
-    = (!eolChar CHAR)
-    /* any CHAR except CR and LF */
-
-/* SequenceItem and set rules */
-
-pub sequence_set -> Vec<SequenceItem>
-    = a:(seq_range / seq_number) b:("," s:sequence_set { s })*
-    {
-        let mut seq = vec![a];
-        for set in b.into_iter() {
-            for elem in set.into_iter() {
-                seq.push(elem);
-            }
-        };
-        seq
-    }
-
-seq_range -> SequenceItem
-    = a:seq_number ":" b:seq_number
-    { Range(Box::new(a), Box::new(b)) }
-
-seq_number -> SequenceItem
-    = (num:nz_number { Number(num) }) / ("*" { Wildcard })
-
-/* RFC 3501 BOILERPLATE */
-
-number -> usize
-    = d:$(DIGIT+)
-    { d.parse().unwrap() }
-    /*
-     * unsigned 32-bit integer
-     * (0 <= n < 4,294,967,296)
-     */
-
-nz_number -> usize
-    = d:$(digit_nz DIGIT*)
-    { d.parse().unwrap() }
-    /*
-     * non-zero unsigned 32-bit integer
-     * (0 < n < 4,294,967,296)
-     */
-
-digit_nz
-    = [1-9]
-
-DIGIT
-    = [0-9]
-
-CHAR
-    = .
-    /* any ASCII char */
-
-CHAR8
-    = (![\x00] CHAR)
-    /* any octet except NUL, \x00 */
-
-DQUOTE
-    = "\""
-
-SP
-    = whitespace
-
-CRLF
-    = eol
-
-/* PEG boilerplate from kevinmehall/rust-peg follows */
-
-__ = (whitespace / eol / comment)*
-
-/* Modeled after ECMA-262, 5th ed., 7.4. */
-comment
-    = singleLineComment
-    / multiLineComment
-
-singleLineComment
-    = "//" (!eolChar .)*
-
-multiLineComment
-    = "/*" (!"*/" .)* "*/"
-
-/* Modeled after ECMA-262, 5th ed., 7.3. */
-eol
-    = "\n"
-    / "\r\n"
-    / "\r"
-    / "\u{2028}"
-    / "\u{2029}"
-
-eolChar
-    = [\n\r\u{2028}\u{2029}]
-    /* Modeled after ECMA-262, 5th ed., 7.2. */
-
-whitespace
-    = [ \t\u{00A0}\u{FEFF}\u{1680}\u{180E}\u{2000}-\u{200A}\u{202F}\u{205F}\u{3000}] // \v\f removed
